Revert "make sure the default tenant is also deleted"
This reverts commit 1633860d97da262b697db21b10eed02029aff12d.

+++ bonita-integration-tests-sp/bonita-integration-tests-client/src/main/java/com/bonitasoft/engine/BPMTestSPUtil.java
@@ -99,11 +99,15 @@ public class BPMTestSPUtil {
         }
         for (final Tenant tenant : tenants) {
             try {
-                platformAPI.deactiveTenant(tenant.getId());
+                if (!tenant.isDefaultTenant()) {
+                    platformAPI.deactiveTenant(tenant.getId());
+                }
             } catch (final TenantDeactivationException tde) {
                 // Nothing to do
             }
-            platformAPI.deleteTenant(tenant.getId());
+            if (!tenant.isDefaultTenant()) {
+                platformAPI.deleteTenant(tenant.getId());
+            }
         }
         platformAPI.stopNode();
         platformAPI.cleanPlatform();
